fix: 스터디 관리 페이지
